Zomato api project/requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
Pillow==8.4.0
